move epoch dataset
