+++ src/levanter/data/text.py
@@ -41,7 +41,7 @@ import haliax as hax
 from haliax import Axis
 
 from levanter.data import AsyncDataset
-from levanter.data.dataset import MappedAsyncDataset
+from levanter.data.dataset import EpochDataset, MappedAsyncDataset
 from levanter.data.mixture import MixtureDataset, StopStrategy, rescale_mixture_schedule_for_batch_schedule
 from levanter.models.lm_model import LmExample
 from levanter.schedule import BatchSchedule
@@ -85,87 +85,6 @@ logger = logging.getLogger("levanter.data.text")
 LEDGER_FILE = "ledger.json"
 
 DEFAULT_IGNORE_INDEX = -100  # Mirrors pytorch's default ignore index
-
-
-class EpochDataset(AsyncDataset[T_co]):
-    """
-    A dataset that wraps another dataset, providing infinite epochs by recycling indices.
-    If `max_epochs` is specified, it limits the number of cycles before raising StopIteration.
-
-    :param dataset: The dataset to wrap.
-    :param max_epochs: The maximum number of epochs to cycle through. If None, cycle indefinitely.
-    """
-
-    def __init__(self, dataset: AsyncDataset[T_co], max_epochs: Optional[int] = None):
-        super().__init__()
-        self.dataset = dataset
-        self.max_epochs = max_epochs
-
-    async def async_len(self) -> int:
-        if self.max_epochs is None:
-            raise ValueError("Cannot determine length of an infinite dataset without max_epochs.")
-        # Return the total number of samples: max_epochs * length of the dataset
-        return self.max_epochs * await self.dataset.async_len()
-
-    async def final_length_is_known(self) -> bool:
-        return await self.dataset.final_length_is_known()
-
-    def is_finite(self) -> bool:
-        # EpochDataset can be finite if max_epochs is set.
-        return self.max_epochs is not None
-
-    async def current_len(self) -> Optional[int]:
-        # If max_epochs is None, the dataset is effectively infinite.
-        if self.max_epochs is None:
-            return None
-
-        # If the final length of the dataset is not known, return the current length of the underlying dataset.
-        if not await self.dataset.final_length_is_known():
-            return await self.dataset.current_len()
-
-        # If the final length is known, return the max_epochs * async_len of the dataset.
-        return self.max_epochs * await self.dataset.async_len()
-
-    async def get_batch(self, indices: Sequence[int]) -> Sequence[T_co]:
-        # Use self.wait_until_len_at_least to ensure we have enough data for the batch.
-        max_index = max(indices)
-        ds_len = await self.dataset.wait_until_len_at_least(max_index + 1)
-
-        # Determine the epoch based on the largest index
-        epoch = max_index // ds_len
-
-        # If max_epochs is specified, raise an error if the epoch exceeds the allowed number of epochs
-        if self.max_epochs is not None and epoch >= self.max_epochs:
-            raise StopIteration(
-                f"Reached maximum number of epochs: epoch {epoch} exceeds the maximum allowed {self.max_epochs}"
-            )
-
-        # Wrap the indices within the bounds of the dataset length
-        wrapped_indices = [idx % ds_len for idx in indices]
-
-        # Delegate to the underlying dataset's get_batch
-        return await self.dataset.get_batch(wrapped_indices)
-
-    async def wait_until_len_at_least(self, length: int) -> int:
-        """
-        Returns the length of the dataset once it is at least `length` or if the dataset has a known (finished) length.
-        If the dataset's actual length is less than `length`, it returns the minimum of async_len and the current length.
-        """
-        # Wait until the underlying dataset's length is at least `length`
-        if not self.is_finite():
-            return length
-
-        if await self.dataset.final_length_is_known():
-            base_length = await self.dataset.async_len()
-        else:
-            base_length = await self.dataset.wait_until_len_at_least(length)
-
-        if base_length < length:
-            # hit epoch boundary
-            assert self.max_epochs is not None
-            return self.max_epochs * base_length
-
-        return base_length
 
 
 class TokenSeqDataset(AsyncDataset[np.ndarray]):
